"No reprieve"
Not sure what I think about her saying "No reprieve" but I'll go with it for now. A major theme of the story is actually forgiveness (particularly forgiveness of ones' self). I didn't want to be so on the nose as to say "No forgiveness." I also considered her not saying anything aloud, or a wording like:

```
No reprieve. It was her rule. It was necessary.
```

That had a nice repetition of "It was," but I do that kind of repetition too often. Also, I didn't want to say "It was her rule"; I want to demonstrate it by having her make the decision at a few more times during this chapter (and definitely in others) to not grant reprieve (for example, later in the chapter there's things like "but she had a rule. She had to." and then she kills one of her (helpless) attackers.