Create Chapter 2 — Dressara.md
This first draft of the beginning of chapter 2 is a bit shitty. Wording very clumsy, a bit too heavy on exposition (it would be better to allow this exposition to come out over a bit more time, spread between more action, so it feels less heavy, IMO)

Still, I have to start this chapter somewhere.